# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: JianArtifact API
  version: 0.4.0
  description: 自托管多格式制品仓库的管理 API。协议端点（Raw/Maven/npm 等原生客户端交互）以协议兼容测试保障，不在此契约内描述。
  license:
    name: MIT
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/repositories:
    get:
      operationId: listRepositories
      summary: 仓库列表（分页，按可见性 / ACL 过滤）
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: 仓库列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createRepository
      summary: 创建仓库
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepositoryRequest"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/repositories/{name}:
    patch:
      operationId: updateRepository
      summary: 更新仓库（可见性 / 配置）
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRepositoryRequest"
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteRepository
      summary: 删除仓库
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
      responses:
        "204":
          description: 已删除
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/repositories/{name}/acl:
    get:
      operationId: getRepositoryAcl
      summary: 读取仓库 ACL
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
      responses:
        "200":
          description: ACL 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AclList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: setRepositoryAcl
      summary: 覆盖写入仓库 ACL
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutAclRequest"
      responses:
        "200":
          description: 已写入
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AclList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/repositories/{name}/online:
    put:
      operationId: setRepositoryOnline
      summary: 设置仓库 online/offline 状态（仅管理员）
      description: |
        仅管理员可操作。online 是节点本地运维状态：不参与复制（M-2），
        对端复制应用仓库变更不会覆盖本地 online，SetOnline 亦不产生复制变更日志。
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRepositoryOnlineRequest"
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/repositories/{name}/recheck-connection:
    post:
      operationId: recheckRepositoryConnection
      summary: 手动重测仓库上游连接（仅管理员）
      description: |
        仅管理员可操作；仅 online 的 proxy 仓库可重测。立即发起一次上游 HEAD 探测，
        更新 auto-block 状态并返回最新连接状态（不等待自动阻止窗口）。
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
      responses:
        "200":
          description: 重测完成，返回最新连接状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionStatus"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/repositories/{name}/assets:
    get:
      operationId: listRepositoryAssets
      summary: 列出仓库制品（分页，可按路径前缀过滤）
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/PrefixParam"
      responses:
        "200":
          description: 制品列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/repositories/{name}/usage:
    get:
      operationId: getRepositoryUsage
      summary: 仓库客户端使用片段（据 format/type 返回接入命令）
      parameters:
        - $ref: "#/components/parameters/RepoNameParam"
      responses:
        "200":
          description: 使用片段
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    RepoNameParam:
      name: name
      in: path
      required: true
      schema:
        type: string
...
    PrefixParam:
      name: prefix
      in: query
      required: false
      description: 按制品路径前缀过滤
      schema:
        type: string
...
  responses:
    BadRequest:
      description: 参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 越权
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        operationId:
          type: string
          description: 统一制品操作失败时返回的追踪标识
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    Repository:
      type: object
      required: [id, name, format, type, visibility, createdAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        format:
          type: string
          enum: [raw, maven, npm, docker, cargo, pypi, gomod, nuget]
        type:
          type: string
          enum: [hosted, proxy, group]
        visibility:
          type: string
          enum: [public, private]
        online:
          type: boolean
          description: 仓库是否在线（管理员可手动置离线；offline 不参与复制）
        connectionStatus:
          $ref: "#/components/schemas/ConnectionStatus"
          description: 上游连接状态（仅 proxy/group 返回；可选字段，向后兼容）
        description:
          type: string
          description: 仓库描述（管理后台可配置，详情页展示）
        remoteUrl:
          type: string
          description: proxy 仓库的上游地址（仅 type=proxy）
        credentialRef:
          type: string
          pattern: "^[A-Z][A-Z0-9_]{0,62}$"
          description: proxy 上游凭据的受限逻辑名称；运行时仅从 JIAN_UPSTREAM_CREDENTIAL_<名称> 读取（仅 type=proxy，非密钥明文）
        immutableRelease:
          type: boolean
          description: hosted 仓库是否拒绝覆盖已存在的 Release
        members:
          type: array
          items:
            type: string
          description: group 仓库的成员仓库名（有序，仅 type=group）
        createdAt:
          type: string
        artifactCount:
          type: integer
          description: 仓库内制品数量（只读统计字段）
        totalSize:
          type: integer
          format: int64
          description: 仓库内制品总字节数（只读统计字段）
    RepositoryList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Repository"
        total:
          type: integer
    CreateRepositoryRequest:
      type: object
      required: [name, format, type]
      properties:
        name:
          type: string
          minLength: 1
        format:
          type: string
          enum: [raw, maven, npm, docker, cargo, pypi, gomod, nuget]
        type:
          type: string
          enum: [hosted, proxy, group]
        visibility:
          type: string
          enum: [public, private]
          default: private
        description:
          type: string
          description: 仓库描述（可选）
        remoteUrl:
          type: string
          description: proxy 仓库的上游地址（type=proxy 时必填）
        credentialRef:
          type: string
          pattern: "^[A-Z][A-Z0-9_]{0,62}$"
          description: proxy 上游凭据的受限逻辑名称；运行时仅从 JIAN_UPSTREAM_CREDENTIAL_<名称> 读取（仅 type=proxy，非密钥明文）
        immutableRelease:
          type: boolean
          description: 创建 hosted 仓库时是否拒绝覆盖已存在的 Release
        members:
          type: array
          items:
            type: string
          description: group 仓库的成员仓库名（有序，type=group 时必填）
    UpdateRepositoryRequest:
      type: object
      properties:
        visibility:
          type: string
          enum: [public, private]
        description:
          type: string
          description: 更新仓库描述（传空串表示清空）
        remoteUrl:
          type: string
          description: 更新 proxy 上游地址（仅 type=proxy）
        credentialRef:
          type: string
          pattern: "^[A-Z][A-Z0-9_]{0,62}$"
          description: 更新 proxy 上游凭据的受限逻辑名称；运行时仅从 JIAN_UPSTREAM_CREDENTIAL_<名称> 读取（仅 type=proxy，非密钥明文）
        immutableRelease:
          type: boolean
          description: 更新 hosted 仓库是否拒绝覆盖已存在的 Release
        members:
          type: array
          items:
            type: string
          description: 更新 group 成员仓库名（仅 type=group）
    SetRepositoryOnlineRequest:
      type: object
      properties:
        online:
          type: boolean
          description: 是否在线（true=在线，false=离线；必填）
    ConnectionStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [READY, AVAILABLE, UNAVAILABLE, AUTO_BLOCKED, OFFLINE]
          description: 上游连接状态（READY=尚未探测，AVAILABLE=可用，AUTO_BLOCKED=自动阻止，UNAVAILABLE=不可用，OFFLINE=手动离线）
        blockedUntil:
          type: string
          format: date-time
          nullable: true
          description: 自动阻止窗口截止时间（非阻止状态为 null）
        description:
          type: string
          description: 状态说明（供展示）
    AclEntry:
      type: object
      required: [subjectId, action]
      properties:
        subjectId:
          type: integer
          format: int64
        action:
          type: string
          enum: [read, write, admin]
    AclList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AclEntry"
    PutAclRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AclEntry"
    AssetSummary:
      type: object
      required: [path, size, hash, updatedAt]
      properties:
        path:
          type: string
        size:
          type: integer
          format: int64
        hash:
          type: string
          description: 内容寻址 blob 的 sha256 摘要
        sha1:
          type: string
          description: 制品内容的 SHA-1 摘要（写入时计算登记）
        md5:
          type: string
          description: 制品内容的 MD5 摘要（写入时计算登记）
        contentType:
          type: string
        createdAt:
          type: string
          description: 首次写入时间
        updatedAt:
          type: string
    AssetList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AssetSummary"
        total:
          type: integer
...
    UsageSnippet:
      type: object
      required: [title, code]
      properties:
        title:
          type: string
        description:
          type: string
        code:
          type: string
    UsageInfo:
      type: object
      required: [format, type, snippets]
      properties:
        format:
          type: string
        type:
          type: string
        description:
          type: string
          description: 仓库描述（匿名详情页页头展示）
        snippets:
          type: array
          items:
            $ref: "#/components/schemas/UsageSnippet"
...